tvision-jsc-1.1.tgz (for IDA 4.7)
Fully implemented keyboard support. Includes updates from Datarescue's Level I release.

+++ source/drivers.cpp
@@ -168,7 +168,7 @@ static const ushort arrowCodes[] =
     kbEnd,  kbDel,   kbIns,kbPgUp, kbPgDn, kbBack
     };
 
-    for( int i = 0; i < sizeof(ctrlCodes); i++ )
+    for( size_t i = 0; i < sizeof(ctrlCodes); i++ )
         if( (keyCode & 0x00ff) == ctrlCodes[i] )
             return arrowCodes[i];
     return keyCode;
